fix: delegate command execution of ignored commands using CommandResolver
